redid styles for hero build added new sieghart pics and pets for grandiel sieghart refactored hero build component
