+++ src/data.js
@@ -111,6 +111,8 @@ import gnar from './assets/pets/gnar.png'
 import goldporkcard from './assets/pets/goldporkcard.png'
 import holy from './assets/pets/holy.png'
 import hoya from './assets/pets/hoya.png'
+import hunt from './assets/pets/hunt.png'
+import luce from './assets/pets/luce.png'
 import luri from './assets/pets/luri.png'
 import manbo from './assets/pets/manbo.png'
 import montero from './assets/pets/montero.png'
@@ -385,7 +387,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr, 
@@ -423,7 +424,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -461,7 +461,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -499,7 +498,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -537,7 +535,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order, 
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: badr, 
     tn1: 'BADR',
     trait2: sadr,
@@ -575,7 +572,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: badr,
@@ -613,7 +609,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -651,7 +646,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -674,8 +668,8 @@ export const heroBuilds= {
     type: 'assault',
     color: '#C12119',
     classPic: assault,
-    pet: '',
-    petName: '',
+    pet: hunt,
+    petName: 'Hunt',
     set: 'Power of Anger (blue)', 
     setColor: '#4E28D0',
     pic: sieghart,
@@ -689,7 +683,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -727,7 +720,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -765,7 +757,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -803,7 +794,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -841,7 +831,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -876,10 +865,9 @@ export const heroBuilds= {
     ring: gale,
     ringName: 'Gale Ring',
     necklace: lava,
-    necklace: 'Lava Necklace',
+    necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -917,7 +905,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -955,7 +942,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -993,7 +979,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -1031,7 +1016,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -1069,7 +1053,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -1107,7 +1090,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -1145,7 +1127,6 @@ export const heroBuilds= {
     necklaceName: 'Lava Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: basi,
@@ -1183,7 +1164,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr, 
@@ -1206,8 +1186,8 @@ export const heroBuilds= {
     type: 'healer',
     color: '#FFE228',
     classPic: healer,
-    pet: '',
-    petName: '',
+    pet: luce,
+    petName: 'Luce',
     set: 'Power of Anger (blue)', 
     setColor: '#4E28D0',
     pic: grandiel,
@@ -1221,7 +1201,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: sadr,
@@ -1259,7 +1238,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: cdr,
@@ -1297,7 +1275,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklace',
     earrings: gust,
     earringsName: 'Earrings of Gust',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: basi,
@@ -1335,7 +1312,6 @@ export const heroBuilds= {
     necklaceName: 'Lifeblood Necklace',
     earrings: order,
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: crit, 
     tn1: 'Crit',
     trait2: sadi,
@@ -1373,7 +1349,6 @@ export const heroBuilds= {
     necklaceName: 'Chasm Necklaces',
     earrings: order, 
     earringsName: 'Earrings of Order',
-    traits: '4 points in each',
     trait1: cdr, 
     tn1: 'CDR',
     trait2: badr,
